requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-redis==5.4.0
djangorestframework==3.15.2
docker==7.1.0
idna==3.10
psycopg2==2.9.10
psycopg2-binary==2.9.10
redis==5.2.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
